+++ CHANGELOG.md
@@ -2,3 +2,6 @@ v0.0.1
 ======
 
  * initial release
+ * lambda arena added
+ * elo competitor added
+ * glicko competitor added